requirements.txt
aiohttp==2.3.10
asn1crypto==0.24.0
async-timeout==2.0.1
asyncio==3.4.3
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
dialogflow==0.5.2
future==0.17.1
google-api-core==1.7.0
google-auth==1.6.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
yarl==1.2.0
